requirements.txt
# Core ML dependencies
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.24.0
scipy>=1.10.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: ARM deployment
# torch-audio>=2.3.0  # For audio processing if needed
# onnx>=1.14.0        # For ONNX export if needed
# onnxruntime>=1.15.0 # For ONNX inference if needed
